💚  fix: Create Dashboard component to manage user goals and integrate GoalForm
